total workload per day is now basis for burndown charts

+++ Mnd.Domain/GraphData.cs
@@ -7,7 +7,7 @@ namespace Mnd.Domain
 {
     public class XYPoint
     {
-        public int X;
+        public double X;
         public double Y;
     }
